优化页面，修复syscode系统代码bug，新增syscode接口

+++ src/main/java/com/ecit/dto/SysCodeDto.java
@@ -13,7 +13,7 @@ import lombok.experimental.Accessors;
 @AllArgsConstructor
 public class SysCodeDto {
 
-    private Integer codeId;
+    private Long codeId;
     private String code;
     private String description;
     private String value;
